Make Server abstract and create WebMcpServer class
