+++ src/tools/contentExploration/searchContent.test.ts
@@ -1,6 +1,6 @@
 import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 
-import { Server } from '../../server.js';
+import { WebMcpServer } from '../../server.web.js';
 import invariant from '../../utils/invariant.js';
 import { Provider } from '../../utils/provider.js';
 import { getMockRequestHandlerExtra } from '../toolContext.mock.js';
@@ -103,7 +103,7 @@ describe('searchContentTool', () => {
   });
 
   it('should create a tool instance with correct properties', () => {
-    const searchContentTool = getSearchContentTool(new Server());
+    const searchContentTool = getSearchContentTool(new WebMcpServer());
     expect(searchContentTool.name).toBe('search-content');
     expect(searchContentTool.description).toContain('searches across all supported content types');
     expect(searchContentTool.paramsSchema).toMatchObject({
@@ -346,7 +346,7 @@ describe('searchContentTool', () => {
 });
 
 async function getToolResult(params: any): Promise<CallToolResult> {
-  const searchContentTool = getSearchContentTool(new Server());
+  const searchContentTool = getSearchContentTool(new WebMcpServer());
   const callback = await Provider.from(searchContentTool.callback);
   return await callback(params, getMockRequestHandlerExtra());
 }
