+++ src/tools/pulse/listMetricsFromMetricIds/listPulseMetricsFromMetricIds.test.ts
@@ -2,7 +2,7 @@ import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 import { Ok } from 'ts-results-es';
 
 import { PulseDisabledError, PulseNotAvailableError } from '../../../errors/mcpToolError.js';
-import { Server } from '../../../server.js';
+import { WebMcpServer } from '../../../server.web.js';
 import invariant from '../../../utils/invariant.js';
 import { Provider } from '../../../utils/provider.js';
 import { getMockRequestHandlerExtra } from '../../toolContext.mock.js';
@@ -32,7 +32,7 @@ describe('listPulseMetricsFromMetricIdsTool', () => {
   });
 
   it('should create a tool instance with correct properties', () => {
-    const listPulseMetricsFromMetricIdsTool = getListPulseMetricsFromMetricIdsTool(new Server());
+    const listPulseMetricsFromMetricIdsTool = getListPulseMetricsFromMetricIdsTool(new WebMcpServer());
     expect(listPulseMetricsFromMetricIdsTool.name).toBe('list-pulse-metrics-from-metric-ids');
     expect(listPulseMetricsFromMetricIdsTool.description).toContain(
       'Retrieves a list of published Pulse Metrics from a list of metric IDs',
@@ -106,7 +106,7 @@ describe('listPulseMetricsFromMetricIdsTool', () => {
 });
 
 async function getToolResult(params: { metricIds: string[] }): Promise<CallToolResult> {
-  const listPulseMetricsFromMetricIdsTool = getListPulseMetricsFromMetricIdsTool(new Server());
+  const listPulseMetricsFromMetricIdsTool = getListPulseMetricsFromMetricIdsTool(new WebMcpServer());
   const callback = await Provider.from(listPulseMetricsFromMetricIdsTool.callback);
   return await callback(params, getMockRequestHandlerExtra());
 }
